added user model and routes, and refractoring
