+++ models/User.js
@@ -1,0 +1,20 @@
+import mongoose from "mongoose"
+import crypto from "crypto"
+
+const userSchema = new mongoose.Schema({
+  userName: {
+    type: String,
+    unique: true,
+    required: true
+  },
+  password: {
+    type: String,
+    required: true
+  },
+  accessToken: {
+    type: String,
+    default: () => crypto.randomBytes(128).toString("hex")
+  }
+})
+
+export const User = mongoose.model("User", userSchema)
